+++ src/pages/MyPage/DeletePage/DeletePage.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Heading from '../../../components/design/Heading/Heading';
+import Heading from '../../../components/common/Heading/Heading';
 import { loginContainer } from '../../LoginPage/LoginPage.style';
 import DeleteForm from './sections/DeleteForm/DeleteForm';
 import WarningMessage from './sections/WarningMessage/WarningMessage';
