+++ src/pages/NotFoundPage/NotFoundPage.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Heading from '@/components/design/Heading/Heading';
+import Heading from '@/components/common/Heading/Heading';
 import {
   notfoundContainer,
   notfoundDescriptStyling,
